fix(cmd): share sanitizeDBName across build tags (beads-nyfc)
intent(cmd-init): keep no-cgo lint and hook runs from failing on init.go's sanitizeDBName reference
decision(build-tags): move sanitizeDBName into an untagged helper file and move its unit test beside it so cgo and no-cgo builds compile the same symbol
constraint(testing): leave cgo-specific store migration tests in store_factory_test.go while the shared helper test lives outside build tags
